+++ locals-naming.tf
@@ -1,6 +1,7 @@
 locals {
-  name_prefix  = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]$/", "$0-") : ""
-  default_name = lower("${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}")
+  # Naming locals/constants
+  name_prefix = lower(var.name_prefix)
+  name_suffix = lower(var.name_suffix)
 
-  autoscale_setting_name = coalesce(var.custom_autoscale_setting_name, "${local.default_name}-autoscale")
+  autoscale_setting_name = coalesce(var.custom_autoscale_setting_name, azurecaf_name.autoscale.result)
 }