+++ variables-naming.tf
@@ -1,0 +1,25 @@
+# Generic naming variables
+variable "name_prefix" {
+  description = "Optional prefix for the generated name"
+  type        = string
+  default     = ""
+}
+
+variable "name_suffix" {
+  description = "Optional suffix for the generated name"
+  type        = string
+  default     = ""
+}
+
+variable "use_caf_naming" {
+  description = "Use the Azure CAF naming provider to generate default resource name. `custom_autoscale_setting_name` override this if set. Legacy default name is used if this is set to `false`."
+  type        = bool
+  default     = true
+}
+
+# Custom naming override
+variable "custom_autoscale_setting_name" {
+  description = "Custom Autoscale setting name"
+  type        = string
+  default     = ""
+}