+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+Breaking
+  * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
+
 Fixed
   * AZ-589: Avoid plan drift when specifying Diagnostic Settings categories